requirements.txt
aiohttp==3.9.5
brotli==1.1.0
brotlicffi==1.1.0.0
ca_certs_locater==1.0
cffi==1.16.0
chardet==5.2.0
cryptography==42.0.4
Django==3.0.5
google_auth_oauthlib==1.2.0
grpc_status==1.0.0
h2==4.1.0
ipython==8.12.3
Jinja2==3.1.3
js==1.0
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
packaging==24.0
pandas==2.2.2
paypalrestsdk==1.13.3
Pillow==10.3.0
pyodide==0.0.2
pyOpenSSL==24.0.0
pyOpenSSL==24.1.0
pyu2f==0.1.5
railroad==0.5.0
scikit_learn==1.4.1.post1
simplejson==3.19.2
skimage==0.0
tensorflow==2.16.1
torch==2.2.0
typing_extensions==4.11.0
xhtml2pdf==0.2.15
zstandard==0.22.0
Django==3.0.5
django-widget-tweaks==1.4.8
sqlparse==0.3.1
social-auth-app-django
paypalrestsdk~=1.13.3
django-paypal
tensorflow~=2.16.1
scikit-learn~=1.4.1.post1
scikit-image~=0.22.0
IPython~=8.21.0
tensorflow
django-paypal
social-auth-app-django
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tensorflow_intel==2.16.1
